some motor tests

+++ include/motors.h
@@ -12,6 +12,6 @@ void dc_motors_init();
 void dc_motors_move(const uint8_t direction, const float power);
 void dc_motors_halt();
 
-void dc_motors_move_for_ms(const uint8_t direction, const uint8_t power, const double time);
+void dc_motors_move_for_ms(const uint8_t direction, const float power, const double time);
 
 #endif
